Additional Compiler Warnings (#73)

+++ src/ipc/broad_phase/voxel_size_heuristic.cpp
@@ -86,7 +86,7 @@ double mean_edge_length(
 
     double sum = 0;
     for (int i = 0; i < edges.rows(); i++) {
-        const size_t e0i = edges(i, 0), e1i = edges(i, 1);
+        const int e0i = edges(i, 0), e1i = edges(i, 1);
         sum += (vertices_t0.row(e0i) - vertices_t0.row(e1i)).norm();
         sum += (vertices_t1.row(e0i) - vertices_t1.row(e1i)).norm();
     }
@@ -94,7 +94,7 @@ double mean_edge_length(
 
     std_deviation = 0;
     for (int i = 0; i < edges.rows(); i++) {
-        const size_t e0i = edges(i, 0), e1i = edges(i, 1);
+        const int e0i = edges(i, 0), e1i = edges(i, 1);
         std_deviation += std::pow(
             (vertices_t0.row(e0i) - vertices_t0.row(e1i)).norm() - mean, 2);
         std_deviation += std::pow(
